:umbrella: Add some more mildly unreviewed LLM-generated unit tests
